build: Move most .gradle files to /gradle and restructure dependencies.gradle
To have a more clean root folder structure moving all .gradle plugin
files (except settings and build.gradle) to /gradle.

Also merged the dependencies.gralde into the build.gralde file, as this
is where most people would expect the dependencies definition. With that
also introduced a small change to the structure of our dependencies
definition:

Previously there was a big list of variables defining the versions of
the dependencies at the top of the dependencies definition. This made it
very hard to determine which specific dependency was meant, because the
variables where for instance named "jsonassert_version", for the
org.skyscreamer.jsonassert dependency. This made it very hard to follow,
which dependencies are being referenced. In order to simplify that, the
variables have been moved into the blocks where the dependencies are
defined and also only variables are used, if multiple dependencies
should always be referenced in the same version. For single dependencies
the version is now directly defined in the respective dependency
statement.