refactor: Remove broadcasting from DataRequest/simplify LoggerManagerVerticle
See NeonBee/architecture/#18. In a recent change a "broadcasting" option
was added to DataRequest and DataVerticle, mainly to publish messages to
every listening verticle on the eventbus. The reason this was added was
due to a newly introduced "LoggerManagerVerticle", a DataVerticle to
maintain the levels of loggers. In order to be able to change the log
levels of all loggers in a clustered NeonBee instance at once, the
broadcasting option could be used to set the log levels to all cluster
nodes.

The introduction of the broadcasting option however, caused multiple
breaks in the concept of inter DataVerticle communication: With the
broadcasting parameter of DataRequest in place any interface methods
exposed in DataSource and DataSink will, in an undocumented manner, not
function as expected:

For DataSource any data returned by the called verticle will simply end
up in void. The called verticle does NOT know about this and thus will
not be able to react accordingly. Even worse, even the calling verticle
has no information whether and which verticles received the boadcsting
request, thus rendering the full interface useless in my opinion. For
DataSink similarely, the broadcasting event can not be tracked by the
calling verticle and a success / failure cannot be reported by by the
called verticle either.

As this option is rather limited in use and will cause side-effects
easily, it was decided in issue #18 to remove the parameter, to maybe be
reintroduced later with a better concept. This however requried to
re-implement parts of the mentioned LoggerManagerVerticle:

It was decided, that instead of using the DataVerticle implementation of
the LoggerManagerVerticle to handle the distribution of the logger
level, the LoggerManagerVerticle will be reduced to a pure JSON-based
DataVerticle which will publish a EventBus message to all cluster nodes
applying a LoggerConfiguration object.

Before this change LoggerConfiguration was much like a POJO with three
values: the logger name, a configuration level and an effective level.
The effective level however, unlike the name suggests, was not the
effective level of the underlying logger, but could have been "null" or
even a different value than the "real" effective logger value. This was
confusing. To overcome this, LoggerConfiguration was changed to
represent only the logger name and a configuration level. The change can
then be applied to the logger by calling a new "applyConfiguredLevel"
method, which will set the effective level. Also calling the
set/getEffectiveLevel methods will now really retrieve or set the level
of the associated logger, instead of just storing another value. This
again simplified the LoggerManagerVerticle implementation.

Last, it was decided to remove the LoggerConfigurations class, which was
previously used for inter-verticle communcation. Instead of using the
LoggerConfigurations class a default List / JsonArray was used, which
should allow for the LoggerManagerVerticle to be more easily used in an
upcoming admin UI implementation for NeonBee. This again reduced the
amount of testable code in NeonBee.