docs: Add CONTRIBUTING.md
Following the convention of many other open source projects, including Vert.x a CONTRIBUTING.md describes the process contributing to our open source project. The guide should clarify how to interact with out project.

As a template the UI5 Tooling [1] CONTRIBUTING.md was taken as a reference. As another open source SAP library, they provided already some good recommendations how processes can be treated, like signing off contributions with the CLA assistant (which still needs to be set-up).

RFCs for now are moved into the "/docs" folder. We may think of splitting this to a separate repository again in future.

[1] https://github.com/SAP/ui5-tooling/blob/master/CONTRIBUTING.md