fix: Broken HTTP status code propagation in ODataEndpointHandler
If the error code of an exception is in the range from 100 to 999, the
ErrorHandler is responsible for setting the HTTP status code. Setting
the status code to -1 signals the ErrorHandler that it must set a
proper status code, based on the passed exception.

This logic was broken, because in Vert.x 4.0 the behavior of failing a
routing context has changed. The new default status code has changed
from -1 to 500.