refactor: simplify the EntityCollectionProcessor
This patch ...

- ... renames getEntities(...) to applyFilterQueryOption(...) because
getEntities(...) was misleading and inconsistent.

- ... removes isDebugEnabled(), because nothing was computed inside.

- ... changes all apply...QueryOption method signatures to throw an
error instead of failing the passed processPromise. This is required,
because if the processPromise has been failed the processing of the
request must be stopped.

- ... removes the count logic from readEntityCollection(...), because it
will be processed in countEntityCollection(...).

- ... removes applySelectQueryOption(...), because it only builds a
select list which should be done at a more central place. This change is
also a prerequisite for the expand query option, which also needs to
modify the select list.

- ... simplifies the countEntityCollection(...) method dramatically.

- ... rename EntityCollectionProcessor into
CountEntityCollectionProcessor, because we implement the
CountEntityCollectionProcessor interface.