refactor: expose SLF4J Logger interface in LoggingFacade
Similar to other examples in NeonBee, NeonBee is trying to make the best
/ state of the art choice, when it comes to be choosing certain
underlying frameworks. When it comes to a logging facade and backend,
SLF4J / Logback was chosen by NeonBee. Previously NeonBee had its own
implementation for the LoggingFacade, possibly making wrong assumtions
when converting arguments into an array before passing it down to the
masqueraded logger.

This change reduces the logging facade to be making as less as possible
assumptions and clearly indicating that the underlying logger follows
the SLF4J rules, by exposing its interface. The JavaDoc was changed to
indicate that masquarading the logger will loose its ability to deal
with markers, also indicated by a UnsupportedOperationExeption thrown
when anyways trying.

As with the pre Vert.x 4 implementation the Vert.x logger did in fact
reorder arguments before forwarding it to the underlying logger
implementation, the plain SLF4J logger and LoggingFacade is no longer
reording arguments. Thus also NeonBee has to follow the SLF4J convention
[1] of putting the throwable as a last parameter into the format list.
This change adapts all calls to any longer.

[1] http://www.slf4j.org/faq.html#paramException