chore: add PMD rules to avoid unnecessary modifiers in tests
JUnit5 is more tolerant regarding the visibilities of Test classes and
methods than JUnit4, which required everything to be public. JUnit5
supports default package, public and protected visibility, even if it is
recommended to use the default package visibility, which improves the
readability of code.

These rules enforce the recommended standard.

Co-authored-by: Pascal Krause <pascal.krause@sap.com>