refactor!: move files to prepare endpoints change
For milestone E1_MS06 configurable endpoints, all logic for populating the server verticle with endpoints should be moved to individual packages and made configurable. This requires to move many files. In order for Git to be able to record the history on said files, the move has been made in this individual commit, breaking code and as a breaking change.