refactor!: move files to prepare endpoints change
For milestone E1_MS06 configurable endpoints, all logic for populating
the server verticle with endpoints should be moved to individual
packages and made configurable. This requires to move many files.

Moving and modifying the same file in a commit, could lead to the
situation that git diff [1] does not detect the move. This could be
prevented by splitting movement and modification into two separate
commits.

This commit won't build and is only valid in combination with its
following partner commit, which represents the modification part.

[1] https://git-scm.com/docs/git-diff#Documentation/git-diff.txt---find-renamesltngt