feat: make launcher options configurable via env
For deployment on cloud environments, it is beneficial to provide the
launcher options via environment variables instead of command line
arguments. This change supports reading options either from command line
or environment variable.

The `withEnvironment` method is a newly introduced helper method which
is useful when certain code should be executed in a context that exports
a set of environment variables.