build(ci): disable clustered tests on Github
The NeonBeeExtension test base is not longer working on Github, because
for unknown reasons the Hazlecast cluster can't longer find reliable its
members on the Github Actions environment. This is only a workaround
until NeonBeeExtension is replaced with a FakeClusterTestBase like the
FakeClusterManager [1] in Vert.x.

Due to the missing cluster tests the test coverage is reduced by 0.5%.

[1] https://github.com/eclipse-vertx/vert.x/blob/master/src/test/java/io/vertx/test/fakecluster/FakeClusterManager.java