feat: allow multiple MANIFST.MF files when parsing NeonBee-Module
Normally there is exactly one MANIFEST.MF on the classpath, because when
parsing a NeonBee-Module a complete isolated classloader is used. But at
the end it doesn't matter how many MANIFEST.MF files are existing as
long as there is only one that contains the NeonBee-Module attributes.
Therefore this change adds more tolerance to the NeonBee-Module
validation.

This also helps in case that a third party library (like Dynatrace)
manipulates the classpath of every URLClassloader by adding own JARs.