refactor: move error handler configuration to ServerConfig class
Before this change the JsonObject passed to the ServerVerticle was used to read the configuration for the custom error template to use. With this change the JsonObject was replaced by the ServerConfig object. In addition to a custom errorHandlerClassName a new errorHandlerTemplate property was added, which can be used to just specify a different (HTML) template to use in the DefaultErrorHandler.

Switched to promisified interfaces in FileSystemHelper. Increased the code coverage error threshold to 80% instead of 81%, as the violation plugin did fail at 81.0% eventhough the failIfLessThanThresholdError was set.