fix: switch to Vert.x-owned Hazelcast instance creation
Previously we initialized a new Hazelcast instance in our own code and passed it to Vert.x. This initialization scheme signals to Vert.x, that we‘d like to be responsible to also take care of shutting down, which we didn‘t want to be in the first place. This commit chages the behaviour so that Vert.x feels responsible to shut down the Hazelcast instance on its own.