feat: add doNotWatchFiles option to not watch for file changes
The WatchVerticle monitors given directories and/or files for changes. Currently it is implemented by the DeployerVerticle and the ModelRefreshVerticle in order to react on file changes.

Similar to the ignoreClassPath option a new doNotWatchFiles option was added, so watching for local file changes can be disabled (e.g. in tests). This also prevents the creation of any WatchService threads on Linux.