ci: use better default options for NeonBee in tests
During investigation of build problems the issue was located that in our tests we have had multiple different ways we used to initialize NeonBee sometimes with and sometimes without a associated NeonBee instance:

- NeonBee.create
- registerNeonBeeMock
- NeonBeeExtension/NeonBeeInstanceConfiguration

Each time NeonBee was initialized we used a different set of options. Especially for tests particular set of NeonBeeOptions is important, e.g. not to scan the class path or not starting the WatchVerticles. Because of the different options how to start Vert.x these options have often been neglected. This commit uses the same set of default options for all the different ways a NeonBee instance is created.

Now when creating a request (with the NeonBeeTestBase.createRequest method), in case the server port is specified in the options, it is preferred over the server port in the server verticle configuration file.