ci: always close Vert.x if tests create own instances
Similar to how in the previous commit NeonBee was created in different ways, also a new Vert.x instance was:

- Injected via the VertxExtension (best case),
- Injected via the NeonBee instance of the NeonBeeExtension,
- Via NeonBee.newVertx, or simply with a good old
- Vertx.vertx() call.

Creation is not that bad, but sometimes it was fogotten to close the associated Vert.x instance leading to stale objects, as if the test was executed on the main thread, it could result in a frozen situation where the GitHub workflows never finish, as non-deamon threads had been still running. E.g. when a cluster was started, but never closed, due to vertx.close() never beeing called. In other cases, eventhough the Vert.x was injected, it was closed by the tests, which led to undeterministic side effects. This commit fixes all instances where a new Vert.x is created.