feat: add custom micrometer registries via NeonBeeConfig
With this change the config parameter micrometerRegistries is
introduced which allows the user to specify a list of (full qualified) class
names, which must implement the load() method of the functional interface
MicrometerRegistryLoader. The load method must return a
MeterRegistry, which will be added to the micrometer registries by
NeonBee. The PrometheusMeterRegistry will be the default
registry, which is only available when the MetricsEndpoint is loaded.

Co-authored-by: Michael Halberstadt <michael.halberstadt@sap.com>