ci: automate dependency upgrade (of vertx)
With the dependency upgrade workflow, bumping the version of vertx (and
in the future also other dependencies) will be automated.

The workflow checks every day at midnight if a new release version of
vertx is available. If it finds a new version, a pull request will be
created which updates the version in the build.gradle.