refactor: improve wildcard handling in SelfFirstClassLoader
So far the SelfFirstClassLoader only allowed to specify parentPreferred className if the wildcard was at the end of the string like:

io.neonbee.*

If the wildcard was in any other position, the check failed. In additional the general implementation of the check was implemented inefficiently, because of a list traversal, on every check, which is, for modules, every class that is to be loaded.

This change improves the wildcard logic, by compiling a pattern on initialization of the SelfFirstClassLoader, also allowing more complex patterns like:

io.neonbee.*.TestClass