feat: add health checks
In order to provide a generic and extendable way of creating health
checks in NeonBee, a central registry is needed where health checks can
be provided. To achieve this, a `HealthCheckRegistry` is added to
NeonBee which will keep track of all default health checks of NeonBee
and custom health checks provided by the NeonBee user.

Vert.x already comes with a health check plugin which will be used for
the health check objects that can be registered on to the registry.
However, as the plugin does not prevent overriding of already existing
health check procedures, a wrapper is necessary to provide this
functionality. In addition, a caching mechanism is implemented which can
be controlled via the retention time. The retention time is the time
from the last health check until a new health check procedure is
triggered. If the next health check is inside this retention window, the
cached result of the previous health check will be returned.

The health checks can be be globally enabled / disabled and a timeout
for the checks can be set via the NeonBeeConfig.

Resolves #117