feat(health): add health check verticle
With this change a `HealthCheckVerticle` is introduced which which is
registering its name into the shared map of NeonBee. It also implements
the `retrieveData` method, which invokes all health checks that are
registered to the health check registry and returns the results of the
checks.

Retrieving and consolidating the data from all HealthCheckVerticles that
are deployed (and registered in the shared map), is done in the
`HealthCheckRegistry` which now offers a new `createRequests` method.
The method returns a consolidated result of all nodes (including both
global and node-specific checks).

The memory and hazelcast cluster health check naming patterns contained
a `/` which is used by the vertx-health-check extension to differentiate
groups. This causes node-specific check to only contain everything after
the last `/` of the id of a check. Therefore, `.` is now used as a new
separator for health check names.