fix: avoid printing warnings when using NeonBeeMockHelper
In the defaultVertxMock() method of the `NeonBeeMockHelper`,
`VertxInternal` should be mocked instead of `Vertx` as the Vertx mock
does not have a method addCloseHook(). Since the method
`NeonBee.registerCloseHandler()` is accessing `addCloseHook` on the
mocked Vertx, it is unkown and thus fails. This prints a warning to the
console when executing tests that pass the Vertx mock created by
defaultVertxMock(...) to the registerNeonBeeMock(...) method.