feat: make ServerVerticle handler configurable
The `ServerVerticle` attaches request handlers to the route. The request
handlers attached to the route cannot be changed at the moment. This
change makes the request handlers that the `ServerVerticle` attaches to
the route configurable.

A `RoutingHandlerFactory` is used to create an instance of the
`Handler<RoutingContext>` class. In the `ServerConfig`, a list of
`RoutingHandlerFactory` class names is used to specify the
`RoutingHandlerFactory`. The handler created by the
`RoutingHandlerFactory` is then added to the root route. The request
handlers are added in the order specified in the `ServerConfig` list.