fix(health): make health check collection more resilient
Whenever a health check verticle did not answer, a failed future was
returned by the health check registry in the attempt of collecting all
health check data. Especially in the clustered case this might
happen quite often due to restarts caused by a deployment, for instance.
In that case an error would be reported, although all relevent data
might be available.

Therefore, this change makes the data collection more resilient by
respecting that some verticles might not be able to answer.

Once the de-registering of health check verticles from the shared map is
available/implemented, it should be less likely that this situation
happens, but still possible.