fix: getName for dummy EntityVerticle
The EntityVerticle#getName returns the same name for each
EntityVerticle created with the createDummyEntityVerticle method.
Therefore, all created EntityVerticles are listing on the same event
bus address and Vert.x is calling the EntityVerticles in a round-robin
fashion.

To fix this, the getName method for the created EntityVerticle objects
is overridden with Mockito#spy.