feat(health): add status endpoint & support filtering
With this change a `StatusEndpoint` is introduced, which allows to
display health onformation in a non-verbose way. This endpoint is meant
to be public and usable for external system monitoring tools. The
endpoint is mounted at /status and is part of the default endpoints of
NeonBee. The /health endpoint will continue to serve health data in a
verbose way, and it should be considered by the user to protect this
endpoint as it might expose system information that should not be
available to external parties.

In addition, the `HealthCheckHandler` is extended to support health
check based filtering vie the request path of the /health endpoint.