feat: add SelfCleaningRegistry and make Registry interface type safe
This change introduces a SelfCleaningRegistry. The SelfCleaningRegistry knows which values were added by which cluster
node. When a node leaves the cluster, all node-related values are removed from the registry.

This change also changes the Registry interface. Now a Registry is type safe and offers supports registration of
multiple values for the same key.