chore: improve error message when no verticle is registered for an entity type
When a request is made to a verticle which requests data from other verticles,
which due to some issue are not registered in the shared map, it becomes difficult
to understand for which entity type there is no verticle registered. Providing the
FQN of the entity type in the error message, makes it easier to identify for which
entity type there is no verticle registered.