fix: mitigate the PartitionMigratingException
When a node leaves or joins the cluster hazelcast will rebalance the
cluster. If we unregister or register Entity verticles when the
hazelcast cluster is not in a safe state we get
PartitionMigratingException's. This commit attempts to postpone the
registration and unregistration until the cluster is in a safe state.