feat: introduce in-memory caching / request coalescing
This change introduces an in-memory caching concept for DataVerticle, as well as an option to coalesce same requests sent to a verticle in parallel. The new CachingDataVerticle allows to define an arbitrary cacheKey that is used to store responses received by the DataVerticle in-memory. The next time data with the same cacheKey is requested, neither the required data verticles, nor the original requireData method is being invoked. Instead the data is returned straight from the in-memory cache. This also allows for request coalescing, meaning that if multiple requests are sent to the verticle with the same cacheKey at the same time, the verticle will only do one call to requestData and return the data straight from cache for the other request instead.