feat: add jsonMaxStringSize configuration
This change makes the maximum string size that is parsed by Jackson customizable. This is achieved by a new ConfigurableJsonFactory SPI that Vert.x can use instead of the default JacksonCodec.

With upgrading to Vert.x v4.4.2, Jackson v2.15 introduced a maximum string parsing limit of 5 Million characters, which equals about 5MB in size. This was further increased to 10 Million / 10 MB with Jackson v2.15.1. However with NeonBee we would like to allow users to deal with larger payloads if required.

This change further improves exception handling on the Vert.x event bus, as a negative side-effect of the Jackson change was that event-bus messages reaching the maximum JSON size resulted in uncached exceptions and thus event bus message timing out, which is the worst case scenario. This change now adds specific tests for failures in different lifecycle states of a message (e.g. when being deserialized on the retriever side vs. being deserialized on the message return), as well as handles failing messages much more gracefully and with a descriptive error message.